backend/requirements.txt
fastapi
uvicorn
pymongo
praw
openai
apscheduler
python-dotenv
